Improve Aspen security.
Since google apps script allows google.script.run calls on *any* function in Apps Script and since we want to share this with teachers and therefore with our domain, we need to consider any exposed function at risk of exploitation.

That means that we need to add security guards around all calls to the Aspen API ensuring only a teacher can call the API and only for their own courses. We implemented a caching method to avoid spamming the API when doing e.g. multiple API calls in a row to get grades, but that caching method means that calls to PropertiesService are *also* insecure.

To fix this, we create an interface putting all API info and properties service getters/setters wrapped inside a function (aspenInterface). If a malicious user calls our aspenInterface, all they get is our function definitions -- they don't have a way to directly get the access token etc.