Support for mapping multiple assignments to rubrics.
Picking the rubric categories is step 1, but this is not complete. We will need to re-work our storage of maps from aspen to google to take into account 1-to-many relationship as opposed to 1-to-1 we have been assuming before. And then we'll have to update our grade posting so it can post updated rubric grades instead of or in addition to the overall grade.